refactor: deviceDetection and fullscreen to core.
