+++ src/core/deviceDetection.ts
